[김기성] anaunseoList session에 아이디 값의 유무에 따른 처리

+++ team27a/WebContent/anaunseo/anaunseoList.jsp
@@ -17,8 +17,14 @@
 		ArrayList<Anaunseo> list = anaunseoDao.selectAnaunseoList();
 		%>
 		<h1>아나운서 목록</h1>
-		<a href="<%= request.getContextPath() %>/anaunseo/insertAnaunseoForm.jsp">등록</a>
-		
+		<%
+		//session에 아이디값이 있다면 등록버튼 활성화
+		if(session.getAttribute("memberId") != null){
+		%>
+			<a href="<%= request.getContextPath() %>/anaunseo/insertAnaunseoForm.jsp">등록</a>
+		<%
+		}
+		%>
 		<!-- table에 thead와 tbody 삽입 -->
 		<table border="1">
 			<thead>
@@ -26,8 +32,15 @@
 					<th>아나운서 순서</th>
 					<th>아나운서 이름</th>
 					<th>아나운서 나이</th>
-					<th>수정</th>
-					<th>삭제</th>
+					<%
+					//session에 아이디값이 있다면 수정 삭제 테이블 제목 활성화
+					if(session.getAttribute("memberId") != null){
+					%>
+						<th>수정</th>
+						<th>삭제</th>
+					<%
+					}
+					%>
 				</tr>
 			</thead>
 			<tbody>
@@ -39,8 +52,15 @@
 						<td><%= anaunseo.getAnaunseoId() %></td>
 						<td><%= anaunseo.getAnaunseoName() %></td>
 						<td><%= anaunseo.getAnaunseoAge() %></td>
+						<%
+						//session에 아이디값이 있다면 수정, 삭제버튼 활성화
+						if(session.getAttribute("memberId") != null){
+						%>
 						<td><a href="<%= request.getContextPath() %>/anaunseo/updateAnaunseoForm.jsp?anaunseoId=<%= anaunseo.getAnaunseoId() %>">수정</a></td>
 						<td><a href="<%= request.getContextPath() %>/anaunseo/deleteAnaunseoAction.jsp?anaunseoId=<%= anaunseo.getAnaunseoId() %>">삭제</a></td>
+						<%
+						}
+						%>
 					</tr>
 				<%
 				}
